Day4/terraform-basics/scripts/requirements.txt
google-cloud-storage==2.14.0
google-cloud-bigquery==3.14.1
pandas==2.1.4
numpy==1.26.2 